Remove migration code
