+++ packages/core/echo/echo-pipeline/src/db-host/index.ts
@@ -5,7 +5,6 @@
 export * from './data-service';
 export * from './documents-synchronizer';
 export * from './echo-host';
-export * from './migration';
 export * from './database-root';
 export * from './query-state';
 export * from './query-service';
