+++ packages/core/echo/echo-protocol/src/index.ts
@@ -5,6 +5,5 @@
 export * from './document-structure';
 export * from './reference';
 export * from './space-doc-version';
-export * from './legacy';
 export * from './collection-sync';
 export * from './space-id';
