+++ packages/core/echo/echo-protocol/src/reference.ts
@@ -6,7 +6,6 @@ import { DXN, LOCAL_SPACE_TAG, PublicKey } from '@dxos/keys';
 import { type ObjectId } from '@dxos/protocols';
 import { type Reference as ReferenceProto } from '@dxos/protocols/proto/dxos/echo/model/document';
 
-import type { LegacyEncodedReferenceObject } from './legacy';
 import { createIdFromSpaceKey } from './space-id';
 
 /**
@@ -89,16 +88,3 @@ export const decodeReference = (value: any) => Reference.fromDXN(DXN.parse(value
 
 export const isEncodedReference = (value: any): value is EncodedReference =>
   typeof value === 'object' && value !== null && Object.keys(value).length === 1 && typeof value['/'] === 'string';
-
-export const convertLegacyReference = async (reference: LegacyEncodedReferenceObject): Promise<EncodedReference> => {
-  if (reference.protocol === Reference.TYPE_PROTOCOL) {
-    return encodeReference(Reference.fromLegacyTypename(reference.itemId));
-  }
-  if (!reference.itemId) {
-    throw new Error('Invalid reference');
-  }
-
-  const spaceKey = reference.host;
-  const spaceId = spaceKey != null ? await createIdFromSpaceKey(PublicKey.fromHex(spaceKey)) : undefined;
-  return encodeReference(new Reference(reference.itemId, reference.protocol ?? undefined, spaceId));
-};
