+++ packages/gravity/proto-guard/src/check-snapshot.node.test.ts
@@ -30,7 +30,8 @@ describe('Load client from storage snapshot', () => {
     return testStoragePath;
   };
 
-  test('check if space loads for echo-levelDB-transition snapshot', { timeout: 10_000 }, async () => {
+  // TODO(dmaretskyi): Doesn't work with migration code gone.
+  test.skip('check if space loads for echo-levelDB-transition snapshot', { timeout: 10_000 }, async () => {
     const snapshot = SnapshotsRegistry.getSnapshot('echo-levelDB-transition');
     invariant(snapshot, 'Snapshot not found');
     log.info('Testing snapshot', { snapshot });
