+++ packages/sdk/client-services/src/packlets/spaces/data-space-manager.ts
@@ -15,7 +15,6 @@ import {
   type MemberInfo,
 } from '@dxos/credentials';
 import {
-  convertLegacyReferences,
   findInlineObjectOfType,
   type EchoEdgeReplicator,
   type EchoHost,
@@ -28,13 +27,7 @@ import {
   type SpaceProtocol,
   type SpaceProtocolSession,
 } from '@dxos/echo-pipeline';
-import {
-  LEGACY_TYPE_PROPERTIES,
-  SpaceDocVersion,
-  encodeReference,
-  type ObjectStructure,
-  type SpaceDoc,
-} from '@dxos/echo-protocol';
+import { SpaceDocVersion, encodeReference, type ObjectStructure, type SpaceDoc } from '@dxos/echo-protocol';
 import { TYPE_PROPERTIES, createObjectId, getTypeReference } from '@dxos/echo-schema';
 import type { EdgeConnection, EdgeHttpClient } from '@dxos/edge-client';
 import { writeMessages, type FeedStore } from '@dxos/feed-store';
@@ -269,9 +262,7 @@ export class DataSpaceManager extends Resource {
         return properties?.data?.[DEFAULT_SPACE_KEY] === this._signingContext.identityKey.toHex();
       }
       case SpaceDocVersion.LEGACY: {
-        const convertedDoc = await convertLegacyReferences(space.databaseRoot.docSync()!);
-        const [_, properties] = findInlineObjectOfType(convertedDoc, LEGACY_TYPE_PROPERTIES) ?? [];
-        return properties?.data?.[DEFAULT_SPACE_KEY] === this._signingContext.identityKey.toHex();
+        throw new Error('Legacy space version is not supported');
       }
 
       default:
